fix: 승리 시 인터벌이 끝나지 않는 문제 수정
phaseUpdate에 있던 승리 확인 로직을 cardUse로 옮겨서 해결.
phaseUpdate는 룸에서 작동하는데 룸안에서 룸을 지운다는 모순이 발생해서 인듯하다.
